Refactor workflow processing to use durable objects for better concurrency
